backend/requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
psycopg2-binary==2.9.11
yfinance==1.2.0
feedparser==6.0.12
requests==2.32.5
pydantic==2.12.5
anyio==4.12.1
